fix: move PR tempalte settings to project preferences
