+++ apps/desktop/src/lib/settings/GithubIntegration.svelte
@@ -1,13 +1,7 @@
 <script lang="ts">
 	import { checkAuthStatus, initDeviceOauth } from '$lib/backend/github';
 	import SectionCard from '$lib/components/SectionCard.svelte';
-	import {
-		gitHostPullRequestTemplatePath,
-		gitHostUsePullRequestTemplate
-	} from '$lib/config/config';
 	import { getGitHubUserServiceStore } from '$lib/gitHost/github/githubUserService';
-	import TextBox from '$lib/shared/TextBox.svelte';
-	import Toggle from '$lib/shared/Toggle.svelte';
 	import { UserService } from '$lib/stores/user';
 	import { copyToClipboard } from '$lib/utils/clipboard';
 	import { getContext } from '$lib/utils/context';
@@ -21,8 +15,6 @@
 	export let minimal = false;
 	export let disabled = false;
 
-	const usePullRequestTemplate = gitHostUsePullRequestTemplate();
-	const pullRequestTemplatePath = gitHostPullRequestTemplatePath();
 	const githubUserService = getGitHubUserServiceStore();
 	const userService = getContext(UserService);
 	const user = userService.user;
@@ -110,32 +102,6 @@
 			<Button style="pop" kind="solid" {disabled} onclick={gitHubStartOauth}>Authorize</Button>
 		{/if}
 	</SectionCard>
-	<div>
-		<SectionCard roundedBottom={false} orientation="row" labelFor="use-pull-request-template">
-			<svelte:fragment slot="title">Pull Request Template</svelte:fragment>
-			<svelte:fragment slot="caption">
-				Use Pull Request template when creating a Pull Requests.
-			</svelte:fragment>
-			<svelte:fragment slot="actions">
-				<Toggle
-					id="use-pull-request-template"
-					value="false"
-					bind:checked={$usePullRequestTemplate}
-				/>
-			</svelte:fragment>
-		</SectionCard>
-		<SectionCard roundedTop={false} orientation="row" labelFor="pull-request-template-path">
-			<svelte:fragment slot="title">Pull Request Template Path</svelte:fragment>
-			<svelte:fragment slot="caption">
-				<div class="pr-path--label">Path to your Pull Request template in your repository.</div>
-				<TextBox
-					id="pull-request-template-path"
-					bind:value={$pullRequestTemplatePath}
-					placeholder=".github/pull_request_template.md"
-				/>
-			</svelte:fragment>
-		</SectionCard>
-	</div>
 {/if}
 
 <Modal
@@ -306,10 +272,6 @@
 		&::before {
 			top: 30px;
 		}
-	}
-
-	.pr-path--label {
-		margin-bottom: 0.75rem;
 	}
 
 	.icon-wrapper {
